backend/requirements.txt
Flask
pytesseract
Pillow